changed: move norm class from bdf to base
