+++ AdvectionDiffusionExplicit.C
@@ -13,7 +13,6 @@
 //==============================================================================
 
 #include "AdvectionDiffusionExplicit.h"
-#include "AdvectionDiffusionBDF.h"
 #include "FiniteElement.h"
